use size_t for size params when sending, int can be negative and is not guaranteed to be large enough

+++ src/common/fe.hpp
@@ -53,7 +53,7 @@ int fe_timeout_add(int interval, GSourceFunc callback, void *userdata);
 void fe_timeout_remove (int tag);
 void fe_new_window (struct session *sess, int focus);
 void fe_new_server (struct server *serv);
-void fe_add_rawlog (struct server *serv, const char *text, int len, int outbound);
+void fe_add_rawlog (struct server *serv, const char *text, size_t len, int outbound);
 #define FE_MSG_WAIT 1
 #define FE_MSG_INFO 2
 #define FE_MSG_WARN 4
